Split out installTheme
